app/requirements/production.txt
Django
django-allauth
django-crispy-forms
django-hitcount
gunicorn
six
pillow